Add ghost karts + item policy option

+++ src/items/powerup.cpp
@@ -631,7 +631,7 @@ void Powerup::hitBonusBox(const ItemState &item_state)
 		int leader_section = item_policy->m_leader_section;
 		if (leader_section == -1) leader_section = 0;
 		ItemPolicySection *curr_sec = &item_policy->m_policy_sections[leader_section];
-		uint16_t rules = curr_sec->m_rules;
+		uint32_t rules = curr_sec->m_rules;
 		bool item_override = rules & ItemPolicyRules::IPT_BONUS_BOX_OVERRIDE;
 
         // We don't check for buckets if full random mode is enabled in config
